MemoryMonitors Implemented 6 kinds of MemoryMonitors that can watch a region in memory for changes and raise events when data in that region is changed.
Non-generic monitors that send raw byte arrays in the EventArgs, generic monitors that marshal data into a specific type, and generic array monitors that watch an array of objects and raises events for individual elements.

Monitors can used from a permanent address or from a pointer plus offset that points to the region of memory that you want to monitor.